+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- configuration with defining achievements model with award_for methods
+
 ## [0.1.0] - 2022-11-30
 ### Added
 - Start project by extracting code from FantasySports project